// src/api/response.type.ts
export interface BaseListType {
  label: string
  value: string | number
}
// ...
export interface ReferralRebateSummary {
  calculate_type: number
  profit: string
  valid_bet_amount: string
  revenue_amount: string
}

export interface ReferralRebateStatementsList {
  list: ReferralRebateStatements[]
  pagination: {
    page: number
    offset: number
    size: number
    total: number
  }
}

export interface ReferralRebateStatements {
  account: string
  tier: number
  rate: number
  currency_id: number
  valid_bet_amount: string
  game_type: number
  deposit_amount: string
  profit: string
  revenue_amount: string
}

export interface ReferralRebateEventsList {
  list: ReferralRebateEvents[]
  pagination: {
    page: number
    offset: number
    size: number
    total: number
  }
}

export interface ReferralRebateEvents {
  id: number
  settlement_time: string
  currency_id: number
  valid_bet_amount: string
  game_type: number
  profit: string
  revenue_amount: string
  distribution_time: string
}